requirements.txt
flask==2.3.3
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0
schedule==1.2.0
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
openai==0.28.1
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
geopy==2.4.0
yfinance==0.2.22
newsapi-python==0.2.6
python-dateutil==2.8.2
sqlalchemy==2.0.21
flask-sqlalchemy==3.0.5
